requirements.txt
google-generativeai
PyQt6
Pillow